backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.41
alembic==1.15.2
asyncpg==0.30.0
psycopg[binary]==3.2.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.9.1
python-multipart==0.0.20
redis==6.0.0
openpyxl==3.1.5
reportlab==4.4.0
httpx==0.28.1
pytest==8.3.5
pytest-asyncio==0.26.0
